# docs/api-contract.openapi.yaml
openapi: "3.1.0"
info:
  title: "Club Informatique de l’IST : contrat d’API"
  version: "1.0.0"
  description: |
    Contrat de référence entre le frontend et le backend.

    Conventions : dates en ISO 8601 UTC ; pagination uniforme (« content », « page », « size », « totalElements », « totalPages ») ;
    erreurs au format RFC 9457 avec un code stable et des messages en français ; identifiants numériques stables.
    Le jeton d’accès est transmis par l’en-tête « Authorization » ; le jeton de rafraîchissement par un cookie HttpOnly.
    L’extension « x-roles » liste les rôles autorisés ; un Formateur et un Responsable du Club héritent des droits du Membre, un Super Admin de ceux de l’Administrateur.
    Les droits sont appliqués par le serveur, par rôle et par propriété de la ressource.
servers:
  - url: /api/v1
...
security:
  - jetonAcces: []
paths:
...
  "/actualites":
    get:
      tags: [Contenus publics]
      summary: Actualités publiées et publiques.
      operationId: getActualites
      security: []
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/size"
        - $ref: "#/components/parameters/sort"
        - $ref: "#/components/parameters/search"
        - $ref: "#/components/parameters/categorieId"
      responses:
        "200":
          description: Succès.
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/PageActualite"
        "400":
          $ref: "#/components/responses/Erreur400"
    post:
      tags: [Gestion du club]
      summary: Créer une actualité.
      operationId: postActualites
      x-roles: [RESPONSABLE_CLUB]
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/ActualiteSaisie"
      responses:
        "201":
          description: Ressource créée.
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Actualite"
        "400":
          $ref: "#/components/responses/Erreur400"
        "401":
          $ref: "#/components/responses/Erreur401"
        "403":
          $ref: "#/components/responses/Erreur403"
  "/actualites/slug/{slug}":
    get:
      tags: [Contenus publics]
      summary: Actualité publiée.
      operationId: getActualitesSlugSlug
      security: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Succès.
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Actualite"
        "404":
          $ref: "#/components/responses/Erreur404"
...
  "/publications":
    get:
      tags: [Espace Membre]
      summary: Annonces publiées réservées aux membres (décision D-01).
      operationId: getPublications
      x-roles: [MEMBRE]
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/size"
        - $ref: "#/components/parameters/sort"
      responses:
        "200":
          description: Succès.
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/PageActualite"
        "400":
          $ref: "#/components/responses/Erreur400"
        "401":
          $ref: "#/components/responses/Erreur401"
        "403":
          $ref: "#/components/responses/Erreur403"
  "/publications/slug/{slug}":
    get:
      tags: [Espace Membre]
      summary: Annonce réservée aux membres.
      operationId: getPublicationsSlugSlug
      x-roles: [MEMBRE]
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Succès.
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Actualite"
        "401":
          $ref: "#/components/responses/Erreur401"
        "403":
          $ref: "#/components/responses/Erreur403"
        "404":
          $ref: "#/components/responses/Erreur404"
...
  "/gestion/actualites":
    get:
      tags: [Gestion du club]
      summary: Toutes les actualités, publiées ou non.
      operationId: getGestionActualites
      x-roles: [RESPONSABLE_CLUB]
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/size"
        - $ref: "#/components/parameters/sort"
        - name: publie
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Succès.
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/PageActualite"
        "400":
          $ref: "#/components/responses/Erreur400"
        "401":
          $ref: "#/components/responses/Erreur401"
        "403":
          $ref: "#/components/responses/Erreur403"
  "/actualites/{id}":
    get:
      tags: [Gestion du club]
      summary: Actualité par identifiant, brouillon compris.
      operationId: getActualitesId
      x-roles: [RESPONSABLE_CLUB]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Succès.
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Actualite"
        "401":
          $ref: "#/components/responses/Erreur401"
        "403":
          $ref: "#/components/responses/Erreur403"
        "404":
          $ref: "#/components/responses/Erreur404"
    put:
      tags: [Gestion du club]
      summary: Modifier une actualité.
      operationId: putActualitesId
      x-roles: [RESPONSABLE_CLUB]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/ActualiteSaisie"
      responses:
        "200":
          description: Succès.
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Actualite"
        "400":
          $ref: "#/components/responses/Erreur400"
        "401":
          $ref: "#/components/responses/Erreur401"
        "403":
          $ref: "#/components/responses/Erreur403"
        "404":
          $ref: "#/components/responses/Erreur404"
    delete:
      tags: [Gestion du club]
      summary: Supprimer une actualité.
      operationId: deleteActualitesId
      x-roles: [RESPONSABLE_CLUB]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Opération réalisée, sans contenu.
        "401":
          $ref: "#/components/responses/Erreur401"
        "403":
          $ref: "#/components/responses/Erreur403"
        "404":
          $ref: "#/components/responses/Erreur404"
  "/actualites/{id}/publication":
    patch:
      tags: [Gestion du club]
      summary: Publier ou retirer une actualité (état explicite).
      operationId: patchActualitesIdPublication
      x-roles: [RESPONSABLE_CLUB]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/Publication"
      responses:
        "200":
          description: Succès.
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Actualite"
        "401":
          $ref: "#/components/responses/Erreur401"
        "403":
          $ref: "#/components/responses/Erreur403"
        "404":
          $ref: "#/components/responses/Erreur404"
...
components:
  securitySchemes:
    jetonAcces:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Numéro de page, à partir de zéro.
    size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 10
      description: Taille de page.
    sort:
      name: sort
      in: query
      schema:
        type: string
      description: "Tri : « champ,asc » ou « champ,desc »."
    search:
      name: search
      in: query
      schema:
        type: string
      description: Recherche plein texte ; facultative.
    categorieId:
      name: categorieId
      in: query
      schema:
        type: integer
        format: int64
  responses:
    Erreur400:
      description: "Requête invalide : erreurs de champ dans « errors »."
      content:
        "application/problem+json":
          schema:
            $ref: "#/components/schemas/Probleme"
    Erreur401:
      description: Session absente ou expirée, ou identifiants refusés.
      content:
        "application/problem+json":
          schema:
            $ref: "#/components/schemas/Probleme"
    Erreur403:
      description: Droits insuffisants (rôle ou propriété de la ressource).
      content:
        "application/problem+json":
          schema:
            $ref: "#/components/schemas/Probleme"
    Erreur404:
      description: Ressource introuvable.
      content:
        "application/problem+json":
          schema:
            $ref: "#/components/schemas/Probleme"
...
  schemas:
    Probleme:
      type: object
      description: Erreur au format RFC 9457. Le champ « code » est stable ; « detail » est rédigé en français.
      required: [title, status, code]
      properties:
        type:
          type: [string, "null"]
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int64
        detail:
          type: [string, "null"]
        instance:
          type: [string, "null"]
        code:
          type: string
        errors:
          type: [array, "null"]
          items:
            $ref: "#/components/schemas/ErreurChamp"
    ErreurChamp:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
...
    Visibilite:
      type: string
      enum: [PUBLIC, MEMBRES]
    Actualite:
      type: object
      required: [id, titre, slug, contenu, publie, visibilite, createdAt]
      properties:
        id:
          type: integer
          format: int64
        titre:
          type: string
        slug:
          type: string
        contenu:
          type: string
        resume:
          type: [string, "null"]
        image:
          type: [string, "null"]
          format: uri
        publie:
          type: boolean
        visibilite:
          $ref: "#/components/schemas/Visibilite"
        datePublication:
          type: [string, "null"]
          format: date-time
        auteurNom:
          type: [string, "null"]
        categorieId:
          type: [integer, "null"]
          format: int64
        categorieNom:
          type: [string, "null"]
        createdAt:
          type: string
          format: date-time
    ActualiteSaisie:
      type: object
      required: [titre, contenu, publie, visibilite]
      properties:
        titre:
          type: string
        contenu:
          type: string
        resume:
          type: [string, "null"]
        image:
          type: [string, "null"]
          format: uri
        categorieId:
          type: [integer, "null"]
          format: int64
        publie:
          type: boolean
        visibilite:
          $ref: "#/components/schemas/Visibilite"
    Publication:
      type: object
      required: [publie]
      properties:
        publie:
          type: boolean
...
    PageActualite:
      type: object
      description: Page de résultats (pagination uniforme).
      required: [content, page, size, totalElements, totalPages]
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/Actualite"
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
...